[console] Add configurable scancode vs BIOS keyboard driver
