+++ elks/arch/i86/defconfig
@@ -136,6 +136,7 @@ CONFIG_BLK_DEV_CHAR=y
 CONFIG_CONSOLE_DIRECT=y
 # CONFIG_CONSOLE_BIOS is not set
 # CONFIG_CONSOLE_HEADLESS is not set
+CONFIG_KEYBOARD_SCANCODE=y
 # CONFIG_CONSOLE_SERIAL is not set
 CONFIG_EMUL_ANSI=y
 # CONFIG_EMUL_VT52 is not set
