+++ elks/arch/i86/drivers/char/bioscon.c
@@ -12,18 +12,9 @@
 
 #include <linuxmt/types.h>
 #include <linuxmt/config.h>
-#include <linuxmt/errno.h>
-#include <linuxmt/fcntl.h>
-#include <linuxmt/fs.h>
-#include <linuxmt/kernel.h>
-#include <linuxmt/major.h>
 #include <linuxmt/mm.h>
-#include <linuxmt/timer.h>
-#include <linuxmt/sched.h>
 #include <linuxmt/chqueue.h>
 #include <linuxmt/ntty.h>
-#include <arch/io.h>
-#include <arch/segment.h>
 #include "console.h"
 #include "bioscon-asm.h"
 
@@ -71,7 +62,6 @@ static Console Con[MAX_CONSOLES], *Visible;
 static Console *glock;		/* Which console owns the graphics hardware */
 static int Width, MaxCol, Height, MaxRow;
 static unsigned short int NumConsoles = MAX_CONSOLES;
-static struct timer_list timer;
 static int kraw;
 static int Current_VCminor = 0;
 
@@ -84,69 +74,6 @@ static int Current_VCminor = 0;
 #endif
 
 static void std_char(register Console *, char);
-
-static void kbd_timer(int data);
-/*
- * Restart timer
- */
-static void restart_timer(void)
-{
-    init_timer(&timer);
-    timer.tl_expires = jiffies + 8;
-    timer.tl_function = kbd_timer;
-    add_timer(&timer);
-}
-
-/*
- * Bios Keyboard Decoder
- */
-static void kbd_timer(int data)
-{
-    int dav, extra = 0;
-
-    if ((dav = bios_kbd_poll())) {
-	if (dav & 0xFF)
-	    Console_conin(dav & 0x7F);
-	else {
-	    dav = (dav >> 8) & 0xFF;
-	    if (dav >= 0x3B && dav <= 0x3D) {	/* temp console switch on F1-F3*/
-		Console_set_vc(dav - 0x3B);
-		dav = 0;
-	    }
-	    else if ((dav >= 0x68) && (dav < 0x6B)) {	/* Change VC */
-		Console_set_vc((unsigned)(dav - 0x68));
-		dav = 0;
-	    }
-	    else if (dav >= 0x3B && dav < 0x45)	/* Function keys */
-		dav = dav - 0x3B + 'a';
-	    else {
-		switch(dav) {
-		case 0x48: dav = 'A'; break;	/* up*/
-		case 0x50: dav = 'B'; break;	/* down*/
-		case 0x4d: dav = 'C'; break;	/* right*/
-		case 0x4b: dav = 'D'; break;	/* left*/
-		case 0x47: dav = 'H'; break;	/* home*/
-		case 0x4f: dav = 'F'; break;	/* end*/
-		case 0x49: dav = '5'; extra = '~'; break; /* pgup*/
-		case 0x51: dav = '6'; extra = '~'; break; /* pgdn*/
-		default:   dav = 0;
-		}
-	    }
-	    if (dav) {
-		Console_conin(ESC);
-#ifdef CONFIG_EMUL_ANSI
-		Console_conin('[');
-#endif
-		Console_conin(dav);
-#ifdef CONFIG_EMUL_ANSI
-		if (extra)
-		    Console_conin(extra);
-#endif
-	    }
-	}
-    }
-    restart_timer();
-}
 
 static void PositionCursor(register Console * C)
 {
@@ -276,8 +203,7 @@ void console_init(void)
 	C++;
     }
 
-    enable_irq(1);		/* enable BIOS Keyboard interrupts */
-    restart_timer();
+    kbd_init();
 
     printk("BIOS console %ux%u"TERM_TYPE"(%u virtual consoles)\n",
 	   Width, Height, NumConsoles);
