+++ emu86-disk.config
@@ -101,6 +101,7 @@ CONFIG_BLK_DEV_CHAR=y
 # CONFIG_CONSOLE_DIRECT is not set
 # CONFIG_CONSOLE_BIOS is not set
 CONFIG_CONSOLE_HEADLESS=y
+# CONFIG_KEYBOARD_SCANCODE is not set
 # CONFIG_CONSOLE_SERIAL is not set
 CONFIG_CHAR_DEV_RS=y
 CONFIG_CHAR_DEV_LP=y
