+++ emu86-rom.config
@@ -114,6 +114,7 @@ CONFIG_BLK_DEV_CHAR=y
 # CONFIG_CONSOLE_DIRECT is not set
 # CONFIG_CONSOLE_BIOS is not set
 CONFIG_CONSOLE_HEADLESS=y
+# CONFIG_KEYBOARD_SCANCODE is not set
 # CONFIG_CONSOLE_SERIAL is not set
 # CONFIG_CHAR_DEV_RS is not set
 # CONFIG_CHAR_DEV_LP is not set
